+++ src/main/java/com/nasilk/createcrystallized/block/entity/PropulsiteThrusterEntity.java
@@ -237,7 +237,7 @@ public class PropulsiteThrusterEntity extends BlockEntity implements IHaveGoggle
             }
 
             // Idling; ambiance is nice
-            if (serverLevel.getRandom().nextFloat() < AMBIENT_RATE) {
+            if (serverLevel.getRandom().nextDouble() < AMBIENT_RATE) {
                 if (charge == 0) serverLevel.playSound(
                     null, worldPosition,
                     SoundEvents.AMETHYST_BLOCK_RESONATE, SoundSource.BLOCKS,
